+++ src/main/java/com/example/RSW/controller/UsrWalkCrewController.java
@@ -143,57 +143,7 @@ public class UsrWalkCrewController {
 		}
 	}
 
-	// ✅ 신청자 목록 보기 (크루장만 접근 가능)
-	@GetMapping("/requestList")
-	public String showRequestList(@RequestParam int crewId, HttpServletRequest req, Model model) {
-		Rq rq = (Rq) req.getAttribute("rq");
-		if (!rq.isLogined()) {
-			return "redirect:/usr/member/login?msg=로그인 후 이용해주세요.";
-		}
-
-		int loginedMemberId = rq.getLoginedMemberId();
-		WalkCrew crew = walkCrewService.getCrewById(crewId);
-
-		if (crew.getLeaderId() != loginedMemberId) {
-			return "redirect:/usr/walkCrew/detail/" + crewId + "?msg=해당 페이지에 접근 권한이 없습니다.";
-		}
-
-		// 신청자 리스트 조회
-		List<Map<String, Object>> applicants = walkCrewService.getApplicantsByCrewId(crewId);
-		model.addAttribute("applicants", applicants);
-		model.addAttribute("crewId", crewId);
-
-		return "usr/walkCrew/requestList";
-	}
-
-	// ✅ 신청자 상세 정보 보기
-	@GetMapping("/requestDetail")
-	public String showRequestDetail(@RequestParam int crewId, @RequestParam int memberId, HttpServletRequest req,
-			Model model) {
-		Rq rq = (Rq) req.getAttribute("rq");
-
-		if (!rq.isLogined()) {
-			return "redirect:/usr/member/login?msg=로그인 후 이용해주세요.";
-		}
-
-		// 크루장인지 확인
-		WalkCrew crew = walkCrewService.getCrewById(crewId);
-		if (crew.getLeaderId() != rq.getLoginedMemberId()) {
-			return "redirect:/usr/walkCrew/detail/" + crewId + "?msg=해당 페이지에 접근 권한이 없습니다.";
-		}
-
-		// 신청자 정보 가져오기
-		Map<String, Object> applicant = walkCrewService.getApplicantDetail(crewId, memberId);
-		if (applicant == null) {
-			return "redirect:/usr/walkCrew/requestList?crewId=" + crewId + "&msg=신청자 정보를 찾을 수 없습니다.";
-		}
-
-		model.addAttribute("applicant", applicant);
-		model.addAttribute("crewId", crewId);
-
-		return "usr/walkCrew/requestDetail";
-	}
-
+	
 	// ✅ 특정 시, 구에 해당하는 동 목록 반환 (Ajax)
 	@GetMapping("/getDongs")
 	@ResponseBody
@@ -217,22 +167,6 @@ public class UsrWalkCrewController {
 		return ResultData.from("S-1", "참가 요청을 수락했습니다.");
 	}
 
-	// 메뉴용 공통 데이터
-	@ModelAttribute("crew")
-	public WalkCrew getCrewForMenu(HttpServletRequest req) {
-		Rq rq = (Rq) req.getAttribute("rq");
-
-		if (rq == null || !rq.isLogined())
-			return null;
-
-		int memberId = rq.getLoginedMemberId();
-
-		WalkCrew crew = walkCrewService.getCrewByLeaderId(memberId);
-		if (crew != null)
-			return crew;
-
-		// ✅ 참가자라도 승인된 경우 크루 정보 반환
-		return walkCrewService.getCrewByMemberId(memberId);
-	}
+	
 
 }
