+++ src/main/java/com/example/RSW/service/WalkCrewMemberService.java
@@ -1,0 +1,32 @@
+package com.example.RSW.service;
+
+import java.util.List;
+import java.util.Map;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.stereotype.Service;
+
+import com.example.RSW.repository.BoardRepository;
+import com.example.RSW.repository.DistrictRepository;
+import com.example.RSW.repository.MemberRepository;
+import com.example.RSW.repository.WalkCrewMemberRepository;
+import com.example.RSW.repository.WalkCrewRepository;
+import com.example.RSW.util.Ut;
+import com.example.RSW.vo.District;
+import com.example.RSW.vo.Member;
+import com.example.RSW.vo.ResultData;
+import com.example.RSW.vo.WalkCrew;
+
+@Service
+public class WalkCrewMemberService {
+
+	@Autowired
+	WalkCrewMemberRepository walkCrewMemberRepository;
+
+	public void requestToJoinCrew(int crewId, int memberId) {
+		return WalkCrewMemberRepository.requestToJoinCrew(crewId, memberId);
+
+	}
+
+}
