+++ src/main/resources/mappers/WalkCrewMemberRepository.xml
@@ -1,0 +1,15 @@
+<?xml version="1.0" encoding="UTF-8" ?>
+<!DOCTYPE mapper
+  PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
+  "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
+<mapper
+	namespace="com.example.RSW.repository.WalkCrewMemberRepository">
+
+	<insert id="requestToJoinCrew">
+		INSERT INTO walk_crew_member
+		(crew_id, member_id, status, request_date)
+		VALUES
+		(#{crewId}, #{memberId}, 'PENDING', NOW())
+	</insert>
+
+</mapper>
